Reorganize ui package
