+++ ScopedStorage/app/src/main/java/com/samples/storage/ActionListAdapter.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 The Android Open Source Project
+ * Copyright (C) 2021 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,8 +12,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
-package com.samples.storage.ui
+package com.samples.storage
 
 import android.view.LayoutInflater
 import android.view.View
@@ -23,7 +24,6 @@ import androidx.annotation.IdRes
 import androidx.annotation.StringRes
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
-import com.samples.storage.R
 
 data class Action(@StringRes val nameRes: Int, @IdRes val actionRes: Int)
