+++ ScopedStorage/app/src/main/java/com/samples/storage/mediastore/MediaStoreFragment.kt
@@ -23,8 +23,8 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.samples.storage.R
 import com.samples.storage.databinding.FragmentListBinding
-import com.samples.storage.ui.Action
-import com.samples.storage.ui.ActionListAdapter
+import com.samples.storage.Action
+import com.samples.storage.ActionListAdapter
 
 private val demoList = arrayOf(
     Action(R.string.mediastore_add, R.id.action_mediaStoreFragment_to_addMediaFragment),
